requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
flake8==3.6.0
gunicorn==19.9.0
mccabe==0.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.1
pytz==2018.9
whitenoise==4.1.2
